+++ keras_nlp/applications/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2021 The KerasNLP Authors
+# Copyright 2022 The KerasNLP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,10 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from keras_nlp import applications
-from keras_nlp import layers
-from keras_nlp import metrics
-from keras_nlp import tokenizers
-from keras_nlp import utils
-
-__version__ = "0.3.0"
+from keras_nlp.applications.bert import BertEncoder
+from keras_nlp.applications.bert import BertLanguageModel
+from keras_nlp.applications.bert import BertClassifier
